Modify unittest structure
Move each unittest to it's own target executable and test
