+++ test/meson.build
@@ -1,7 +1,2 @@
 
-unittests_exe = executable( 'unittest', 'unittest.c', ut_sources,
-            include_directories : inc, 
-            link_with : hashtable,
-            c_args : '-DUNIT_TEST' )
-
-test('unit tests', unittests_exe)
+subdir('unittests')
